FrontDesk/agent/requirements.txt
livekit==0.12.1
livekit-api==0.6.0
python-dotenv==1.0.1
aiohttp==3.10.5
openai==1.50.2
requests==2.32.3